chore: move serializers to sinks package
