+++ platform_plugin_aspects/sinks/external_id_sink.py
@@ -1,6 +1,6 @@
 """User profile sink"""
 
-from platform_plugin_aspects.serializers import UserExternalIDSerializer
+from platform_plugin_aspects.sinks.serializers import UserExternalIDSerializer
 from platform_plugin_aspects.sinks.base_sink import ModelBaseSink
