+++ platform_plugin_aspects/sinks/user_profile_sink.py
@@ -1,6 +1,6 @@
 """User profile sink"""
 
-from platform_plugin_aspects.serializers import UserProfileSerializer
+from platform_plugin_aspects.sinks.serializers import UserProfileSerializer
 from platform_plugin_aspects.sinks.base_sink import ModelBaseSink
